requirements.txt
geocoder==1.38.1
mysql-connector-python==8.3.0
pypdf==3.9.0
PyYAML==6.0.1
